requirements.txt
python-telegram-bot==20.8
yt-dlp==2024.3.10
python-dotenv==1.0.0
